requirements.txt
flask
transformers
torch
openpyxl
python-docx
pandas
